refactor(ui): localize context menu policy
